+++ apps/backend/prisma/schema.prisma
@@ -19,7 +19,7 @@ model Category {
 model ProductImage {
   id         String   @id @default(uuid())
   imageUrl   String   @map("image_url")
-  productId  String   @map("productId")
+  productId  String   @map("product_id")
   product    Product  @relation(fields: [productId], references: [id])
   createdAt DateTime  @default(now()) @map("created_at")
   updatedAt DateTime  @updatedAt @map("updated_at")
