+++ apps/backend/src/core/graphql/schema.gql
@@ -3,11 +3,11 @@
 # ------------------------------------------------------
 
 type Category {
-  createdAt: DateTime!
-  id: ID!
-  imageUrl: String
-  name: String!
-  updatedAt: DateTime!
+	createdAt: DateTime!
+	id: ID!
+	imageUrl: String
+	name: String!
+	updatedAt: DateTime!
 }
 
 """
@@ -16,37 +16,37 @@ A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date
 scalar DateTime
 
 type Mutation {
-  createProduct(createProduct: ProductDTO!): Product!
-  createProductImage(createProductImage: ProductImageDTO!): ProductImage!
+	createProduct(createProduct: ProductDTO!): Product!
+	createProductImage(createProductImage: ProductImageDTO!): ProductImage!
 }
 
 type Product {
-  createdAt: DateTime!
-  id: ID!
-  name: String!
-  updatedAt: DateTime!
+	createdAt: DateTime!
+	id: ID!
+	name: String!
+	updatedAt: DateTime!
 }
 
 input ProductDTO {
-  name: String!
+	name: String!
 }
 
 type ProductImage {
-  createdAt: DateTime!
-  id: ID!
-  imageUrl: String!
-  productId: String!
-  updatedAt: DateTime!
+	createdAt: DateTime!
+	id: ID!
+	imageUrl: String!
+	productId: String!
+	updatedAt: DateTime!
 }
 
 input ProductImageDTO {
-  imageUrl: String!
-  productId: String!
+	imageUrl: String!
+	productId: String!
 }
 
 type Query {
-  ProductImage(id: String!): ProductImage
-  ProductImages: [ProductImage!]!
-  categories: [Category!]!
-  category(id: String!): Category
+	categories: [Category!]!
+	category(id: String!): Category
+	productImage(id: String!): ProductImage
+	productImages: [ProductImage!]!
 }
