+++ apps/backend/src/modules/product-image/dto/product-image.input.ts
@@ -1,0 +1,10 @@
+import { Field, InputType } from "@nestjs/graphql";
+
+@InputType()
+export class ProductImageDTO {
+	@Field(() => String)
+	imageUrl: string;
+
+	@Field(() => String)
+	productId: string;
+}
