--- apps/backend/src/modules/product-image/dto/productimage.input.ts
@@ -1,10 +1,0 @@
-import { InputType, Field } from '@nestjs/graphql';
-
-@InputType()
-export class ProductImageDTO {
-    @Field(() => String)
-    imageUrl: string;
-
-    @Field(() => String)
-    productId: string;
-}
