--- apps/backend/src/modules/product-image/entities/productimage.entity.ts
@@ -1,20 +1,0 @@
-import { ObjectType, Field, ID } from '@nestjs/graphql';
-import { Product } from '../../product/entities/product.entity';
-
-@ObjectType()
-export class ProductImage {
-    @Field(() => ID)
-    id: string;
-
-    @Field(() => String)
-    imageUrl: string;
-
-    @Field(() => String)
-    productId: string;
-
-    @Field(() => Date)
-    createdAt: Date;
-
-    @Field(() => Date)
-    updatedAt: Date;
-}
