+++ apps/backend/src/modules/product-image/product-image.module.ts
@@ -1,8 +1,8 @@
-import { Module } from '@nestjs/common';
-import { ProductImageService } from './product-image.service';
-import { ProductImageResolver } from './product-image.resolver';
+import { Module } from "@nestjs/common";
+import { ProductImageResolver } from "./product-image.resolver";
+import { ProductImageService } from "./product-image.service";
 
 @Module({
-  providers: [ProductImageResolver, ProductImageService],
+	providers: [ProductImageResolver, ProductImageService],
 })
 export class ProductImageModule {}
