+++ apps/backend/src/modules/product-image/product-image.resolver.spec.ts
@@ -1,19 +1,19 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ProductImageResolver } from './product-image.resolver';
-import { ProductImageService } from './product-image.service';
+import { Test, TestingModule } from "@nestjs/testing";
+import { ProductImageResolver } from "./product-image.resolver";
+import { ProductImageService } from "./product-image.service";
 
-describe('ProductImageResolver', () => {
-  let resolver: ProductImageResolver;
+describe("ProductImageResolver", () => {
+	let resolver: ProductImageResolver;
 
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [ProductImageResolver, ProductImageService],
-    }).compile();
+	beforeEach(async () => {
+		const module: TestingModule = await Test.createTestingModule({
+			providers: [ProductImageResolver, ProductImageService],
+		}).compile();
 
-    resolver = module.get<ProductImageResolver>(ProductImageResolver);
-  });
+		resolver = module.get<ProductImageResolver>(ProductImageResolver);
+	});
 
-  it('should be defined', () => {
-    expect(resolver).toBeDefined();
-  });
+	it("should be defined", () => {
+		expect(resolver).toBeDefined();
+	});
 });
