+++ apps/backend/src/modules/product-image/product-image.service.spec.ts
@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ProductImageService } from './product-image.service';
+import { Test, TestingModule } from "@nestjs/testing";
+import { ProductImageService } from "./product-image.service";
 
-describe('ProductImageService', () => {
-  let service: ProductImageService;
+describe("ProductImageService", () => {
+	let service: ProductImageService;
 
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [ProductImageService],
-    }).compile();
+	beforeEach(async () => {
+		const module: TestingModule = await Test.createTestingModule({
+			providers: [ProductImageService],
+		}).compile();
 
-    service = module.get<ProductImageService>(ProductImageService);
-  });
+		service = module.get<ProductImageService>(ProductImageService);
+	});
 
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
+	it("should be defined", () => {
+		expect(service).toBeDefined();
+	});
 });
