+++ apps/backend/src/modules/product/product.module.ts
@@ -1,8 +1,8 @@
-import { Module } from '@nestjs/common';
-import { ProductService } from './product.service';
-import { ProductResolver } from './product.resolver';
+import { Module } from "@nestjs/common";
+import { ProductResolver } from "./product.resolver";
+import { ProductService } from "./product.service";
 
 @Module({
-  providers: [ProductResolver, ProductService],
+	providers: [ProductResolver, ProductService],
 })
 export class ProductModule {}
