+++ apps/backend/src/modules/product/product.resolver.spec.ts
@@ -1,19 +1,19 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ProductResolver } from './product.resolver';
-import { ProductService } from './product.service';
+import { Test, TestingModule } from "@nestjs/testing";
+import { ProductResolver } from "./product.resolver";
+import { ProductService } from "./product.service";
 
-describe('ProductResolver', () => {
-  let resolver: ProductResolver;
+describe("ProductResolver", () => {
+	let resolver: ProductResolver;
 
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [ProductResolver, ProductService],
-    }).compile();
+	beforeEach(async () => {
+		const module: TestingModule = await Test.createTestingModule({
+			providers: [ProductResolver, ProductService],
+		}).compile();
 
-    resolver = module.get<ProductResolver>(ProductResolver);
-  });
+		resolver = module.get<ProductResolver>(ProductResolver);
+	});
 
-  it('should be defined', () => {
-    expect(resolver).toBeDefined();
-  });
+	it("should be defined", () => {
+		expect(resolver).toBeDefined();
+	});
 });
