+++ apps/backend/src/modules/product/product.resolver.ts
@@ -1,17 +1,16 @@
-import { Resolver, Mutation, Args, Int } from '@nestjs/graphql';
-import { ProductService } from './product.service';
-import { Product } from './entities/product.entity';
-import { ProductDTO } from './dto/create-product.input';
+import { Args, Int, Mutation, Resolver } from "@nestjs/graphql";
+import { ProductDTO } from "./dto/create-product.input";
+import { Product } from "./entities/product.entity";
+import { ProductService } from "./product.service";
 
 @Resolver(() => Product)
 export class ProductResolver {
-  constructor(private readonly productService: ProductService) {}
+	constructor(private readonly productService: ProductService) {}
 
-  @Mutation(() => Product) 
-  async createProduct(
-    @Args('createProduct') payload: ProductDTO
-  ): Promise<Product> {
-    return await this.productService.createProduct(payload);
-  }
-
+	@Mutation(() => Product)
+	async createProduct(
+		@Args("createProduct") payload: ProductDTO,
+	): Promise<Product> {
+		return await this.productService.createProduct(payload);
+	}
 }
