+++ apps/backend/src/modules/product/product.service.spec.ts
@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ProductService } from './product.service';
+import { Test, TestingModule } from "@nestjs/testing";
+import { ProductService } from "./product.service";
 
-describe('ProductService', () => {
-  let service: ProductService;
+describe("ProductService", () => {
+	let service: ProductService;
 
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [ProductService],
-    }).compile();
+	beforeEach(async () => {
+		const module: TestingModule = await Test.createTestingModule({
+			providers: [ProductService],
+		}).compile();
 
-    service = module.get<ProductService>(ProductService);
-  });
+		service = module.get<ProductService>(ProductService);
+	});
 
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
+	it("should be defined", () => {
+		expect(service).toBeDefined();
+	});
 });
