+++ package.json
@@ -23,6 +23,6 @@
 	"devDependencies": {
 		"@biomejs/biome": "1.9.4",
 		"lefthook": "^1.8.2",
-		"turbo": "^2.3.3"
+		"turbo": "^2.4.1"
 	}
 }
